feat: add `Store::alternate_db_paths()`.
Provides a way to learn about loose database paths that are provided by
git alternates.
