+++ gix-odb/tests/fixtures/generated-archives/.gitignore
@@ -1,1 +1,2 @@
 repo_with_loose_objects.tar.xz
+make_alternates_odb.tar.xz
